Feature/otimizacao integracao front

+++ api/src/main/java/com/AJTBackend/dto/CotacaoResponseDTO.java
@@ -1,10 +1,12 @@
 package com.AJTBackend.dto;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
+// usa BigDecimal (e nao Double) pra nao perder precisao na taxa de cambio
 public record CotacaoResponseDTO(
-        Double amount,
+        BigDecimal amount,
         String base,
         String date,
-        Map<String, Double> rates
+        Map<String, BigDecimal> rates
 ) {}
